moved tests into separated folders
